+++ src/Pay.php
@@ -16,6 +16,7 @@ use Yansongda\Artful\Exception\ContainerException;
 use Yansongda\Artful\Exception\ServiceNotFoundException;
 use Yansongda\Pay\Provider\Airwallex;
 use Yansongda\Pay\Provider\Alipay;
+use Yansongda\Pay\Provider\Allinpay;
 use Yansongda\Pay\Provider\Bestpay;
 use Yansongda\Pay\Provider\Douyin;
 use Yansongda\Pay\Provider\Jsb;
@@ -25,6 +26,7 @@ use Yansongda\Pay\Provider\Unipay;
 use Yansongda\Pay\Provider\Wechat;
 use Yansongda\Pay\Service\AirwallexServiceProvider;
 use Yansongda\Pay\Service\AlipayServiceProvider;
+use Yansongda\Pay\Service\AllinpayServiceProvider;
 use Yansongda\Pay\Service\BestpayServiceProvider;
 use Yansongda\Pay\Service\DouyinServiceProvider;
 use Yansongda\Pay\Service\JsbServiceProvider;
@@ -42,6 +44,7 @@ use Yansongda\Pay\Service\WechatServiceProvider;
  * @method static Douyin    douyin(array<string, mixed> $config = [], $container = null)
  * @method static Paypal    paypal(array<string, mixed> $config = [], $container = null)
  * @method static Stripe    stripe(array<string, mixed> $config = [], $container = null)
+ * @method static Allinpay  allinpay(array<string, mixed> $config = [], $container = null)
  * @method static Bestpay   bestpay(array<string, mixed> $config = [], $container = null)
  */
 class Pay
@@ -57,6 +60,7 @@ class Pay
     public const PROVIDER_DOUYIN = 'douyin';
     public const PROVIDER_PAYPAL = 'paypal';
     public const PROVIDER_STRIPE = 'stripe';
+    public const PROVIDER_ALLINPAY = 'allinpay';
     public const PROVIDER_BESTPAY = 'bestpay';
 
     /**
@@ -86,6 +90,7 @@ class Pay
         DouyinServiceProvider::class,
         PaypalServiceProvider::class,
         StripeServiceProvider::class,
+        AllinpayServiceProvider::class,
         BestpayServiceProvider::class,
     ];