+++ src/Plugin/Allinpay/CallbackPlugin.php
@@ -1,0 +1,71 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Yansongda\Pay\Plugin\Allinpay;
+
+use Closure;
+use Yansongda\Artful\Contract\PluginInterface;
+use Yansongda\Artful\Direction\NoHttpRequestDirection;
+use Yansongda\Artful\Exception\ContainerException;
+use Yansongda\Artful\Exception\InvalidConfigException;
+use Yansongda\Artful\Exception\InvalidParamsException;
+use Yansongda\Artful\Exception\ServiceNotFoundException;
+use Yansongda\Artful\Logger;
+use Yansongda\Artful\Rocket;
+use Yansongda\Pay\Config\AllinpayConfig;
+use Yansongda\Pay\Exception\Exception;
+use Yansongda\Pay\Exception\InvalidSignException;
+use Yansongda\Pay\Pay;
+use Yansongda\Pay\Traits\AllinpayTrait;
+use Yansongda\Supports\Collection;
+
+/**
+ * @see https://prodoc.allinpay.com/doc/191/ 通联交易结果通知（POST form 回调，验签后应答 success）
+ */
+class CallbackPlugin implements PluginInterface
+{
+    use AllinpayTrait;
+
+    /**
+     * @throws ContainerException
+     * @throws InvalidConfigException
+     * @throws InvalidParamsException
+     * @throws ServiceNotFoundException
+     * @throws InvalidSignException
+     */
+    public function assembly(Rocket $rocket, Closure $next): Rocket
+    {
+        Logger::info('[Allinpay][CallbackPlugin] 插件开始装载', ['rocket' => $rocket]);
+
+        $this->formatRequestAndParams($rocket);
+
+        $params = $rocket->getParams();
+
+        /** @var AllinpayConfig $config */
+        $config = self::getProviderConfig(Pay::PROVIDER_ALLINPAY, $params);
+
+        self::verifyAllinpaySign($config, $rocket->getPayload());
+
+        $rocket->setDirection(NoHttpRequestDirection::class)
+            ->setDestination($rocket->getPayload());
+
+        Logger::info('[Allinpay][CallbackPlugin] 插件装载完毕', ['rocket' => $rocket]);
+
+        return $next($rocket);
+    }
+
+    /**
+     * @throws InvalidParamsException
+     */
+    protected function formatRequestAndParams(Rocket $rocket): void
+    {
+        $request = $rocket->getParams()['_request'] ?? null;
+
+        if (!$request instanceof Collection) {
+            throw new InvalidParamsException(Exception::PARAMS_CALLBACK_REQUEST_INVALID);
+        }
+
+        $rocket->setPayload($request)->setParams($rocket->getParams()['_params'] ?? []);
+    }
+}